memcmp, fix to previous commit, size_t not present, use int
git-svn-id: svn://coreboot.org/openfirmware@2354 1552c027-8020-0410-b4b5-a757f869b4ce

+++ clients/lib/mem.c
@@ -17,7 +17,7 @@ memset(char *cp, int c, int len)
 }
 
 int
-memcmp(const void *s1, const void *s2, size_t n)
+memcmp(const void *s1, const void *s2, int n)
 {
 	int diff;
 	while (n--) {
